Rename internal files used by stressRooFit and StressRooStats from .cxx to .h Since these files are included by the main program, cmake detects the dependency only if their suffix is .h and not .cxx
